+ Ability to add local offline artifact repositories + File chooser: support custom descriptions

+++ jipipe-core/src/main/java/org/hkijena/jipipe/api/artifacts/JIPipeArtifactRepositoryInstallArtifactRun.java
@@ -22,9 +22,11 @@ import org.hkijena.jipipe.utils.WebUtils;
 import org.hkijena.jipipe.utils.json.JsonUtils;
 
 import java.io.IOException;
+import java.net.URI;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 
 public class JIPipeArtifactRepositoryInstallArtifactRun extends JIPipeArtifactRepositoryOperationRun {
 
@@ -52,32 +54,56 @@ public class JIPipeArtifactRepositoryInstallArtifactRun extends JIPipeArtifactRe
         }
 
         String suffix = artifact.getUrl().endsWith(".zip") ? ".zip" : ".tar.gz";
-        Path tmpFile = JIPipeRuntimeApplicationSettings.getTemporaryFile("artifact", suffix);
-        try {
-            //Download
-            WebUtils.download(new URL(artifact.getUrl()), tmpFile, "Download", progressInfo);
-            if (progressInfo.isCancelled())
-                return;
 
-            // Extract
-            if (suffix.equals(".zip")) {
-                ArchiveUtils.decompressZipFile(tmpFile, targetPath, progressInfo.resolve("Extracting"));
-            } else {
-                ArchiveUtils.decompressTarGZ(tmpFile, targetPath, progressInfo.resolve("Extracting"));
+        URI uri = URI.create(artifact.getUrl());
+        if ("file".equalsIgnoreCase(uri.getScheme())) {
+            try {
+                Path filePath = Paths.get(uri);
+                progressInfo.log("Extracting local file " + filePath);
+
+                // Extract
+                if (suffix.equals(".zip")) {
+                    ArchiveUtils.decompressZipFile(filePath, targetPath, progressInfo.resolve("Extracting"));
+                } else {
+                    ArchiveUtils.decompressTarGZ(filePath, targetPath, progressInfo.resolve("Extracting"));
+                }
+
+                // Create metadata file
+                if (progressInfo.isCancelled())
+                    return;
+                JsonUtils.saveToFile(artifact, targetPath.resolve("artifact.json"));
+            } catch (IOException e) {
+                throw new RuntimeException(e);
             }
+        }
+        else {
+            Path tmpFile = JIPipeRuntimeApplicationSettings.getTemporaryFile("artifact", suffix);
+            try {
+                //Download
+                WebUtils.download(new URL(artifact.getUrl()), tmpFile, "Download", progressInfo);
+                if (progressInfo.isCancelled())
+                    return;
 
-            // Create metadata file
-            if (progressInfo.isCancelled())
-                return;
-            JsonUtils.saveToFile(artifact, targetPath.resolve("artifact.json"));
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        } finally {
-            if (Files.exists(tmpFile)) {
-                try {
-                    Files.delete(tmpFile);
-                } catch (IOException e) {
-                    progressInfo.log("Warning: could not delete " + tmpFile);
+                // Extract
+                if (suffix.equals(".zip")) {
+                    ArchiveUtils.decompressZipFile(tmpFile, targetPath, progressInfo.resolve("Extracting"));
+                } else {
+                    ArchiveUtils.decompressTarGZ(tmpFile, targetPath, progressInfo.resolve("Extracting"));
+                }
+
+                // Create metadata file
+                if (progressInfo.isCancelled())
+                    return;
+                JsonUtils.saveToFile(artifact, targetPath.resolve("artifact.json"));
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            } finally {
+                if (Files.exists(tmpFile)) {
+                    try {
+                        Files.delete(tmpFile);
+                    } catch (IOException e) {
+                        progressInfo.log("Warning: could not delete " + tmpFile);
+                    }
                 }
             }
         }
